refactor: reorganize model browser into platform structure
