+++ src/platform/models/browser/components/ModelBrowserDialog.vue
@@ -138,17 +138,17 @@ import SearchBox from '@/components/common/SearchBox.vue'
 import VirtualGrid from '@/components/common/VirtualGrid.vue'
 import BaseModalLayout from '@/components/widget/layout/BaseModalLayout.vue'
 import LeftSidePanel from '@/components/widget/panel/LeftSidePanel.vue'
-import { useModelBrowserFiltering } from '@/composables/useModelBrowserFiltering'
-import { useModelFilterOptions } from '@/composables/useModelFilterOptions'
-import { useModelKeyboardNav } from '@/composables/useModelKeyboardNav'
-import { useModelLoader } from '@/composables/useModelLoader'
+import { useModelBrowserFiltering } from '@/platform/models/browser/composables/useModelBrowserFiltering'
+import { useModelFilterOptions } from '@/platform/models/browser/composables/useModelFilterOptions'
+import { useModelKeyboardNav } from '@/platform/models/browser/composables/useModelKeyboardNav'
+import { useModelLoader } from '@/platform/models/browser/composables/useModelLoader'
 import { useModelTypes } from '@/platform/assets/composables/useModelTypes'
 import { useTelemetry } from '@/platform/telemetry'
 import type { ComfyModelDef } from '@/stores/modelStore'
-import type { EnrichedModel } from '@/types/modelBrowserTypes'
+import type { EnrichedModel } from '@/platform/models/browser/types/modelBrowserTypes'
 import type { NavGroupData } from '@/types/navTypes'
 import { OnCloseKey } from '@/types/widgetTypes'
-import { createModelNode } from '@/utils/modelBrowser/createModelNode'
+import { createModelNode } from '@/platform/models/browser/utils/createModelNode'
 
 import ModelBrowserStates from './ModelBrowserStates.vue'
 import ModelCard from './ModelCard.vue'
