+++ src/platform/models/browser/components/ViewModeToggle.vue
