+++ src/platform/models/browser/composables/useModelBrowserFiltering.test.ts
@@ -2,8 +2,8 @@ import { describe, expect, it, vi } from 'vitest'
 import { ref } from 'vue'
 
 import { ComfyModelDef } from '@/stores/modelStore'
-import type { EnrichedModel } from '@/types/modelBrowserTypes'
-import { transformToEnrichedModel } from '@/utils/modelBrowser/modelTransform'
+import type { EnrichedModel } from '@/platform/models/browser/types/modelBrowserTypes'
+import { transformToEnrichedModel } from '@/platform/models/browser/utils/modelTransform'
 
 import { useModelBrowserFiltering } from './useModelBrowserFiltering'
