+++ src/platform/models/browser/composables/useModelLoader.test.ts
