+++ src/platform/models/browser/utils/modelTypeIcons.ts
